backend-python-api/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
httpx        # for calling external APIs (e.g. TMDb, AniList, OpenLibrary, RAWG.io)
pydantic     # for data models